Evaluate the performance of a signal detection procedure

diagnose_detection evaluates the performance of a signal detection procedure comparing the output selection table to a reference selection table

diagnose_detection(reference, detection, by.sound.file = FALSE,
time.diagnostics = FALSE, parallel = 1, pb = TRUE)

Arguments

reference
Data frame or 'selection.table' (following the warbleR package format) with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end". It must contain the reference selections that will be used for detection optimization.

detection
Data frame or 'selection.table' with the detections (start and end of the signals) that will be compared against the 'reference' selections. Must contained at least the following columns: "sound.files", "selec", "start" and "end". It can contain data for additional sound files not found in 'references'. In this case the routine assumes that no signals are found in those files, so detection from those files are all false positives.

by.sound.file
Logical argument to control whether performance diagnostics are summarized across sound files (when by.sound.file = FALSE, when more than 1 sound file is included in 'reference') or shown separated by sound file. Default is FALSE.

time.diagnostics
Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if TRUE). Default is FALSE.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar. Default is TRUE.

Value

A data frame including the following detection performance diagnostics:

true.positives: number of signals in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.

false.positives: number of detections that don't match any of the signals in 'reference'. In a perfect detection routine it should be 0.

false.negatives: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.

split.positives: number of signals in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.

merged.positives: number of signals in 'detection' that were overlapped by more than 1 detection (i.e. signals that were merged). In a perfect detection routine it should be 0.

mean.duration.true.positives: mean duration of true positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.positives: mean duration of false positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.negatives: mean duration of false negatives (in s). Only included when time.diagnostics = TRUE.

proportional.overlap.true.positives: ratio of the time overlap of true positives in 'detection' with its corresponding reference signal to the duration of the reference signal.

proportional.duration.true.positives: ratio of duration of true positives to th duration of signals in 'reference'. In a perfect detection routine it should be 1. Based only on true positives that were not split or merged

sensitivity: Proportion of signals in 'reference' that were detected. In a perfect detection routine it should be 1.

specificity: Proportion of detections that correspond to signals in 'reference' that were detected. In a perfect detection routine it should be 1.

Details

The function evaluates the performance of a signal detection procedure by comparing its output selection table to a reference selection table in which all signals of interest have been selected.

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# perfect detection
diagnose_detection(reference = lbh_selec_reference, detection = lbh_selec_reference)

# missing one in detection
diagnose_detection(reference = lbh_selec_reference, detection = lbh_selec_reference[-1, ])

# an extra one in detection
diagnose_detection(reference = lbh_selec_reference[-1, ], detection = lbh_selec_reference)

# with time diagnostics
diagnose_detection(reference = lbh_selec_reference[-1, ],
detection = lbh_selec_reference, time.diagnostics = TRUE)

# and extra sound file in reference
diagnose_detection(reference = lbh_selec_reference,
detection =
lbh_selec_reference[lbh_selec_reference$sound.files != "Phae.long1.wav", ])

# and extra sound file in detection
diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files != "Phae.long1.wav", ],
detection = lbh_selec_reference)

# and extra sound file in detection by sound file
dd <- diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files != "Phae.long1.wav", ],
detection = lbh_selec_reference, time.diagnostics = TRUE, by.sound.file = TRUE)

# get summary
summarize_diagnostic(dd)

}