Summarize detection diagnostics

summarize_diagnostic summarizes detection diagnostics

summarize_diagnostic(diagnostic, time.diagnostics = FALSE)

Arguments

diagnostic

A data frame with the reference selections (start and end of the signals) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".

time.diagnostics

Logical argument to control if diagnostics related to the duration of the signals ("mean.duration.true.positives", "mean.duration.false.positives", "mean.duration.false.negatives" and "proportional.duration.true.positives") are returned (if TRUE). Default is FALSE.

Value

A data frame, typically the output of a detection optimization function (diagnose_detection, optimize_energy_detector, optimize_template_detector) including the following detection performance diagnostics:

true.positives: number of signals in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.

false.positives: number of detections that don't match any of the signals in 'reference'. In a perfect detection routine it should be 0.

false.negatives: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.

split.positives: number of signals in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.

merged.positives: number of signals in 'detection' that were overlapped by more than 1 detection (i.e. signals that were merged). In a perfect detection routine it should be 0.

mean.duration.true.positives: mean duration of true positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.positives: mean duration of false positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.negatives: mean duration of false negatives (in s). Only included when time.diagnostics = TRUE.

overlap.to.true.positives: ratio of the time overlap of true positives in 'detection' with its corresponding reference signal to the duration of the reference signal.

proportional.duration.true.positives: ratio of duration of true positives to th duration of signals in 'reference'. In a perfect detection routine it should be 1. Based only on true positives that were not split or merged. Only included when time.diagnostics = TRUE.

recall: Proportion of signals in 'reference' that were detected. In a perfect detection routine it should be 1.

precision: Proportion of detections that correspond to signals in 'reference' that were detected. In a perfect detection routine it should be 1.

f1.score: Combines recall and precision as the harmonic mean of these two. Provides a single value for evaluating performance. In a perfect detection routine it should be 1.

Details

The function summarizes a detection diagnostic data frame in which diagnostic parameters are shown split by (typically) a categorical column, usually sound files. This function is used internally by diagnose_detection.

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# load example selection tables

data("lbh_reference")

# run diagnose_detection() by sound file
diag <- diagnose_detection(reference = lbh_reference,
detection = lbh_reference[-1, ], by.sound.file = TRUE)

# summarize
summarize_diagnostic(diagnostic = diag)

# should be the same as this:
diagnose_detection(reference = lbh_reference,
detection = lbh_reference[-1, ], by.sound.file = FALSE)
}