energy_detector detects the start and end of acoustic signals

energy_detector detects the start and end of acoustic signals

energy_detector(files = NULL, envelopes = NULL, path = NULL, wl = 512, power = 1,
thinning = 1, bp = NULL, ssmooth = 0, filter = "ffilter", threshold = 15, hold.time = 0,
min.duration = NULL, max.duration = NULL, parallel = 1, pb = TRUE)

Arguments

files
Character vector indicating the sound files that will be analyzed. Optional. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory.

envelopes
An object of class 'envelopes' (generated by get_envelopes) containing the amplitude envelopes of the sound files to be analyzed. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

wl
A numeric vector of length 1 specifying the window used internally by ffilter for bandpass filtering (so only applied when 'bp' is supplied). Default is 512.

power
A numeric vector of length 1 indicating a power factor applied to the amplitude envelope. Increasing power will reduce low amplitude modulations and increase high amplitude modulations, in order to reduce background noise. Default is 1 (no change). This argument is used internally by get_envelopes. Not used if 'envelopes' are supplied.

thinning
Numeric vector of length 1 in the range 0~1 indicating the proportional reduction of the number of samples used to represent amplitude envelopes (i.e. the thinning of the envelopes). Usually amplitude envelopes have many more samples than those needed to accurately represent amplitude variation in time, which affects the size of the output (usually very large R objects / files). Default is 1 (no thinning). Higher sampling rates can afford higher size reduction (e.g. lower thinning values). Reduction is conducted by interpolation using approx. Note that thinning may decrease time precision, and the higher the thinning the less precise the time detection. This argument is used internally by get_envelopes. Not used if 'envelopes' are supplied.

bp
Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is NULL. This argument is used internally by get_envelopes. Not used if 'envelopes' are supplied.

ssmooth
A numeric vector of length 1 to smooth the amplitude envelope with a sum smooth function. Default is 0 (no smoothing). Note that smoothing is applied before thinning (see 'thinning' argument). This argument is used internally by get_envelopes. Not used if 'envelopes' are supplied. #' @param ssmooth A numeric vector of length 1 to smooth the amplitude envelope with a sum smooth function. Default is 0. Note that smoothing is applied before thinning (see 'thinning' argument).

filter
Character vector of length 1 indicating the bandpass filter to be applied (only used if 'bp' is supplied). Three options available, (corresponding to the frequency filter functions in the 'seewave' package): ffilter (ffilter), bwfilter (bwfilter) and fir (fir). Not used if 'envelopes' are supplied.

threshold
A numeric vector of length 1 specifying the amplitude threshold for detecting signals (in %).

hold.time
Numeric vector of length 1. Specifies the time range at which selections will be merged (i.e. if 2 selections are separated by less than the specified hold.time they will be merged in to a single selection). Default is 0 (no hold time applied).

min.duration
Numeric vector of length 1 giving the shortest duration (in seconds) of the signals to be detected. It removes signals below that threshold. If 'hold.time' is supplied signals are first merged and then filtered by duration.

max.duration
Numeric vector of length 1 giving the longest duration (in seconds) of the signals to be detected. It removes signals above that threshold. If 'hold.time' is supplied signals are first merged and then filtered by duration.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar. Default is TRUE.

Value

A data frame containing the start and end of each signal by sound file. If no signal was detected for a sound file it is not included in the output data frame.

Details

This function determines the start and end of signals in sound files. Sound files should be located in the working directory or the path to the sound files should be provided using the 'path' argument. The routine steps are: 1) calculating/modfiying amplitude envelopes (unless 'envelopes' is supplied), 2) detecting signals above the threshold, 3) merging signals (if 'hold.time' is supplied) and 4) filtering signals based on duration (if 'min.duration' and/or 'max.duration' are supplied). Note that when envelopes are not supplied they are calculated on the fly which is more efficient in terms of memory usage (better option if R crashes!).

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr). Implements a modified version of the timer function from seewave.

Examples

{
 if (FALSE) {
# Save example files into temporary working directory
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# using smoothing and minimum duration
detec <- energy_detector(files = c("Phae.long1.wav", "Phae.long2.wav",
"Phae.long3.wav", "Phae.long4.wav"),
path = tempdir(), threshold = 6, ssmooth = 300,
bp = c(2, 9), wl = 300, min.duration = 0.09)

# diagnose detection
diagnose_detection(reference = lbh_selec_reference,
detection = detec)

# without declaring 'files'
detec <- energy_detector(path = tempdir(), threshold = 6, ssmooth = 300,
bp = c(2, 9), wl = 300, min.duration = 0.09)

# diagnose detection
diagnose_detection(reference = lbh_selec_reference,
detection = detec)

# using hold time
detec <- energy_detector(threshold = 10, hold.time = 0.15,
bp = c(2, 9), wl = 300, path = tempdir())

# diagnose detection
diagnose_detection(reference = lbh_selec_reference, detection = detec)

# calculate envelopes first
envs <- get_envelopes(bp = c(2, 9), wl = 300, path = tempdir())

# then run detection providing 'envelopes' (but no 'files')
detec <- energy_detector(envelopes = envs, threshold = 10, hold.time = 0.15, min.duration = 0.05)

# diagnose detection
diagnose_detection(reference = lbh_selec_reference, detection = detec, time.diagnostics = TRUE)

# USIN OTHER SOUND FILE FORMAT (flac program must be installed)
 # fisrt convert files to flac
 warbleR::wav_2_flac(path = tempdir())

 # change sound file extension to flac
 flac_reference <- lbh_selec_reference
 flac_reference$sound.files <- gsub(".wav", ".flac", flac_reference$sound.files)

 # run detection
 detec <- energy_detector(files = c("Phae.long1.flac", "Phae.long2.flac",
 "Phae.long3.flac", "Phae.long4.flac"), path = tempdir(), threshold = 6,
 ssmooth = 300, bp = c(2, 9), wl = 300, min.duration = 0.09)

 # diagnose detection
 diagnose_detection(reference = flac_reference, detection = detec)
 }
}