Merge overlapping selections

merge_overlaps merges several overlapping selections a single selection

merge_overlaps(X, pb = TRUE, parallel = 1)

Arguments

X
Data frame or 'selection.table' (following the warbleR package format) with selections (start and end of the signals). Must contained at least the following columns: "sound.files", "selec", "start" and "end".

pb
Logical argument to control progress bar. Default is TRUE.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

Value

If any time-overlapping selection is found it returns a data frame in which overlapping selections are collapse into a single selection.

Details

The function finds time-overlapping selection in data frames or selection table and collapses them into a single selection. Useful to prepare reference tables to be used in an energy detection routine. In such cases overlapping selections are expected to be detected as a single sound. Therefore, Merging them in the reference table creates a more realistic expectation of how a pefect energy detection routine would look like.

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# load data
data("lbh_reference")

# nothing to merge
merge_overlaps(lbh_reference)

# create artificial overlapping selections
lbh_ref2 <- rbind(as.data.frame(lbh_reference[c(3, 10), ]), lbh_reference[c(3, 10), ])

lbh_ref2$selec <- 1:nrow(lbh_ref2)

merge_overlaps(lbh_ref2)
}