Optimize energy-based signal detection

Optimize energy-based signal detection under different correlation treshold values

optimize_energy_detector(reference, files = NULL, threshold = 15, power = 1,
 hop.size = 11.6, wl = NULL, smooth = 5, hold.time = 0, min.duration = NULL,
 max.duration = NULL, thinning = 1, parallel = 1, pb = TRUE,
 by.sound.file = FALSE, bp = NULL, path = NULL, previous.output = NULL)

Arguments

reference
Selection table (using the warbleR package's format, see selection_table) or data frame with columns for sound file name (sound.files), selection number (selec), and start and end time of signal (start and end). It must contain the reference selections that will be used for detection optimization.

files
Character vector indicating the sound files that will be analyzed. Optional. If not supplied the function will work on the sound files in 'reference'. It can be used to include signals with no signals.

threshold
A numeric vector specifying the amplitude threshold for detecting signals (in %). Default is 15. Several values can be supplied for optimization.

power
A numeric vector indicating a power factor applied to the amplitude envelope. Increasing power will reduce low amplitude modulations and increase high amplitude modulations, in order to reduce background noise. Default is 1 (no change). Several values can be supplied for optimization.

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram. Default is NULL. If supplied, 'hop.size' is ignored. Used internally for bandpass filtering (so only applied when 'bp' is supplied).

smooth
A numeric vector to smooth the amplitude envelope with a sum smooth function. It controls the time range (in ms) in which amplitude samples are smoothed (i.e. averaged with neighboring samples). Default is 5. 0 means no smoothing is applied. Note that smoothing is applied before thinning (see 'thinning' argument). Several values can be supplied for optimization.

hold.time
Numeric vector of length 1. Specifies the time range at which selections will be merged (i.e. if 2 selections are separated by less than the specified hold.time they will be merged in to a single selection). Default is 0 (no hold time applied). Several values can be supplied for optimization.

min.duration
Numeric vector giving the shortest duration (in seconds) of the signals to be detected. It removes signals below that threshold. Several values can be supplied for optimization.

max.duration
Numeric vector giving the longest duration (in seconds) of the signals to be detected. It removes signals above that threshold. Several values can be supplied for optimization.

thinning
Numeric vector in the range 0~1 indicating the proportional reduction of the number of samples used to represent amplitude envelopes (i.e. the thinning of the envelopes). Usually amplitude envelopes have many more samples than those needed to accurately represent amplitude variation in time, which affects the size of the output (usually very large R objects / files). Default is 1 (no thinning). Higher sampling rates may afford higher size reduction (e.g. lower thinning values). Reduction is conducted by interpolation using approx. Note that thinning may decrease time precision, and the higher the thinning the less precise the time detection. Several values can be supplied for optimization.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control progress bar and messages. Default is TRUE.

by.sound.file
Logical argument to control whether performance diagnostics are summarized across sound files (when by.sound.file = FALSE, when more than 1 sound file is included in 'reference') or shown separated by sound file. Default is FALSE.

bp
Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is NULL. This argument is used internally by get_envelopes. Not used if 'envelopes' are supplied.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

previous.output
Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.

Value

A data frame in which each row shows the result of a detection job with a particular combination of tuning parameters (including in the data frame). It also includes the following diagnostic metrics:

true.positives: number of signals in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.

false.positives: number of detections that don't match any of the signals in 'reference'. In a perfect detection routine it should be 0.

false.negatives: number of signals in 'reference' that were not detected (not found in 'detection'. In a perfect detection routine it should be 0.

split.positives: number of signals in 'reference' that were overlapped by more than 1 detection (i.e. detections that were split). In a perfect detection routine it should be 0.

merged.positives: number of signals in 'detection' that were overlapped by more than 1 detection (i.e. signals that were merged). In a perfect detection routine it should be 0.

mean.duration.true.positives: mean duration of true positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.positives: mean duration of false positives (in s). Only included when time.diagnostics = TRUE.

mean.duration.false.negatives: mean duration of false negatives (in s). Only included when time.diagnostics = TRUE.

proportional.overlap.true.positives: ratio of the time overlap of true positives in 'detection' with its corresponding reference signal to the duration of the reference signal.

proportional.duration.true.positives: ratio of duration of true positives to th duration of signals in 'reference'. In a perfect detection routine it should be 1. Based only on true positives that were not split or merged. Only included when time.diagnostics = TRUE.

sensitivity: Proportion of signals in 'reference' that were detected. In a perfect detection routine it should be 1.

specificity: Proportion of detections that correspond to signals in 'reference' that were detected. In a perfect detection routine it should be 1.

Details

This function takes a selections data frame or 'selection_table' ('reference') estimates the detection performance of a energy detector under different detection parameter combinations. This is done by comparing the position in time of the detection to those of the reference selections in 'reference'. The function returns several diagnostic metrics to allow user to determine which parameter values provide a detection that more closely matches the selections in 'reference'. Those parameters can be later used for performing a more efficient detection using energy_detector.

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr).

Examples

{
# Save example files into temporary working directory
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_reference"))
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))
writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"))

# using smoothing and minimum duration
optimize_energy_detector(reference = lbh_selec_reference, path = tempdir(),
threshold = c(0.06, 0.1), smooth = 6.8, bp = c(2, 9), hop.size = 6.8,
min.duration = 0.09)

# with thinning and smoothing
optimize_energy_detector(reference = lbh_selec_reference, path = tempdir(),
 threshold = c(0.06, 0.1, 0.15), smooth = c(7, 10), thinning = c(0.1, 0.01),
 bp = c(2, 9), hop.size = 6.8, min.duration = 0.09)

# by sound file
(opt_ed <- optimize_energy_detector(reference = lbh_selec_reference,
path = tempdir(), threshold = c(0.06, 0.1, 0.15), smooth = 6.8, bp = c(2, 9),
hop.size = 6.8, min.duration = 0.09, by.sound.file = TRUE))

# summarize
summarize_diagnostic(opt_ed)

# using hold time
(op_ed <- optimize_energy_detector(reference = lbh_selec_reference,
threshold = 0.1, hold.time = c(0.1, 0.15), bp = c(2, 9), hop.size = 6.8,
path = tempdir()))

# including previous output in new call
optimize_energy_detector(reference = lbh_selec_reference, threshold = 0.1,
hold.time = c(0.05, 0.2), previous.output = op_ed, smooth = 6.8,
bp = c(2, 9), hop.size = 7, path = tempdir())

# having and extra file in files (simulating a file that should have no detetions)
sub_reference <- lbh_selec_reference[lbh_selec_reference$sound.files != "Phae.long1.wav", ]

optimize_energy_detector(reference = sub_reference, files = unique(lbh_selec_reference$sound.files),
threshold = 0.1, hold.time = c(0.1, 0.15), bp = c(2, 9), smooth = 6.8,
hop.size = 7, path = tempdir())
}