Acoustic template detection from time-frequency cross-correlations

template_detector find signal occurrences in cross-correlation vectors from template_correlator

template_detector(template.correlations, parallel = 1, threshold, pb = TRUE,
 verbose = TRUE)

Arguments

template.correlations
object of class 'template_correlations' generated by template_correlator containing the correlation score vectors.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

threshold
Numeric vector of length 1 with a value between 0 and 1 specifying the correlation threshold for detecting signal occurrences (i.e. correlation peaks). Must be supplied. Correlation scores are forced to between 0 and 1 (by converting negative scores to 0). 0 and 1 represent the lowest and highest similarity to the template respectively.

pb
Logical argument to control progress bar. Default is TRUE.

verbose
Logical argument to control if some summary messages are printed to the console.

Value

The function returns a data frame with the start and end and correlation score for the detected signals.

Details

This function infers signals occurrences from cross-correlation scores along sound files. Correlation scores must be generated first using template_correlator.

References

Araya-Salas, M. (2021), ohun: automatic detection of acoustic signals. R package version 0.1.0.

See also

Author

Marcelo Araya-Salas marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "lbh_selec_table2", "comp_matrix"))

# save sound files
writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"))
writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"))
writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"))

# template for the first sound file in 'lbh_selec_reference'
templ1 <- lbh_selec_reference[1, ]

# generate template correlations
tc <- template_correlator(templates = templ1, path = tempdir(), files = "Phae.long1.wav")

# template detection
td <- template_detector(template.correlations = tc, threshold = 0.4)

# diagnose detection
diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files == "Phae.long1.wav", ],
detection = td)

# template for the second and third sound file in 'lbh_selec_reference'
# which have similar song types
templ2 <- lbh_selec_reference[4, ]

# generate template correlations
tc <- template_correlator(templates = templ2, path = tempdir(),
files = c("Phae.long1.wav", "Phae.long2.wav"))

# template detection
td <- template_detector(template.correlations = tc, threshold = 0.3)

# diagnose detection
diagnose_detection(reference =
lbh_selec_reference[lbh_selec_reference$sound.files %in% c("Phae.long3.wav",
"Phae.long2.wav"), ],
detection = td)
}